third_party/blink/web_tests/http/tests/encoding/meta-switch-mid-parse-with-title-expected.txt

ALERT: 0x59 0xfd 0x6c 0x6d 0x61 0x7a 0x20 0xd6 0x5a 0x44 0xdd 0x4c 
ALERT: 0x59 0x131 0x6c 0x6d 0x61 0x7a 0x20 0xd6 0x5a 0x44 0x130 0x4c 
This test shows that we fix up the title of the document when the encoding changes mid parse. 
